+++ R/vroom.R
@@ -20,7 +20,7 @@ NULL
 #'   default, no variable will be created.
 #' @param col_select Columns to include in the results. You can use the same
 #'   mini-language as `dplyr::select()` to refer to the columns by name. Use
-#'   `c()` or `list()` to use more than one selection expression. Although this
+#'   `c()` to use more than one selection expression. Although this
 #'   usage is less common, `col_select` also accepts a numeric column index. See
 #'   [`?tidyselect::language`][tidyselect::language] for full details on the
 #'   selection language.
@@ -77,7 +77,7 @@ NULL
 #' vroom(input_file, col_select = starts_with("d"))
 #'
 #' # You can also rename specific columns
-#' vroom(input_file, col_select = list(car = model, everything()))
+#' vroom(input_file, col_select = c(car = model, everything()))
 #'
 #' # Column types --------------------------------------------------------------
 #' # By default, vroom guesses the columns types, looking at 1000 rows